Run nix-build in CI
I think `cachix-action` runs it only when there is push access to the cache. There isn't one in this repo, so I guess that's why it is not running nix-build. In any case, it is better to run it explicitly.

Let's wait and see if the CI for this PR in fact runs nix-build.